+++ scripts/join-assert-docs.mjs
@@ -1,71 +1,83 @@
 import fs from 'fs/promises';
 import path from 'path';
 
+async function extractCommonTranslation() {
+  const fileContent = await fs.readFile(path.join('src', 'translationMap.ts'));
+  const regex = /\/\/@@start([\s\S]*?)\/\/@@end/g;
+  const match = regex.exec(fileContent);
+  if (!match) throw new Error('Something went wrong with Translation map');
+  const contentBetweenTags = match[1].trim();
+  if (!contentBetweenTags) throw new Error('Something went wrong with Translation map');
+
+  const indentMatch = contentBetweenTags.match(/^\s*/);
+  const indent = indentMatch ? indentMatch[0] : '';
+
+  const formattedContent = contentBetweenTags
+    .trim()
+    .split('\n')
+    .map((line) => indent + line.trim())
+    .join('\n');
+  return formattedContent;
+}
+
 function extractJsDocContent(fileContent) {
-  // Regex to match JSDoc comments
   const jsDocRegex = /\/\*\*[\s\S]*?\*\//g;
-  // Find all JSDoc comments in the file content
   const jsDocMatches = fileContent.match(jsDocRegex);
 
   if (!jsDocMatches) {
-    // Return an empty string if no JSDoc comments are found
     return '';
   }
 
-  // Clean up and process each JSDoc comment
   const cleanedJsDocContent = jsDocMatches
     .map((jsDoc) => {
-      // Remove the starting /** and ending */
       let cleanedJsDoc = jsDoc.replace(/^\/\*\*|\*\/$/g, '').trim();
-
-      // Remove leading * from each line
-      // cleanedJsDoc = cleanedJsDoc.replace(/(^|\n)\s*\*\s?/g, '\n').trim();
-
-      // Find @example blocks and wrap them with code fences
       cleanedJsDoc = cleanedJsDoc.replace(/(@example[\s\S]*?)(\n\s*\* @)/g, (match, exampleContent, nextTag) => {
-        // Add TypeScript code block markers
         return `@example\n\`\`\`typescript\n${exampleContent.replace(/^@example\s*/, '')}\`\`\`\n${nextTag}`;
       });
 
       cleanedJsDoc = cleanedJsDoc.replace(/(^|\n)\s*\*/g, '\n').trim();
-
-      cleanedJsDoc = cleanedJsDoc.replace(/@example\s*/, '* _Example_\n').trim();
-
       cleanedJsDoc = cleanedJsDoc.replace(/@description\s*/, '* _Description_ ').trim();
-      cleanedJsDoc = cleanedJsDoc.replace(/@param\s*/, '* _Param_ ').trim();
-      cleanedJsDoc = cleanedJsDoc.replace(/@returns\s*/, '* _Returns_ ').trim();
-      cleanedJsDoc = cleanedJsDoc.replace(/@translation\s*/, '* _See_ ').trim();
-      cleanedJsDoc = cleanedJsDoc.replace(/@throws\s*/, '* _Throws_ ').trim();
-      
+      cleanedJsDoc = cleanedJsDoc.replace(/ @example\s*/, '* _Example_\n').trim();
+      cleanedJsDoc = cleanedJsDoc.replace(/ @param\s*/, '* _Param_ ').trim();
+      cleanedJsDoc = cleanedJsDoc.replace(/ @returns.+\n/, '').trim();
+      cleanedJsDoc = cleanedJsDoc.replace(/ @translation\s*/, '* _See_ ').trim();
+      cleanedJsDoc = cleanedJsDoc.replace(/ @throws\s*/, '* _Throws_ ').trim();
 
       return cleanedJsDoc;
     })
-    .join('\n\n'); // Separate different JSDoc blocks with a double newline
+    .join('\n\n');
 
   return cleanedJsDocContent;
 }
 
 const run = async () => {
-  let textData = '## Custom Assert Documentation \n';
+  let readmeStart = await fs.readFile(path.join('scripts', 'readmeStart.md'), { encoding: 'utf8' });
+  const translationCommonMap = await extractCommonTranslation();
+  readmeStart = readmeStart.replace(/@@TRANSLATION_COMMON_MAP@@/, translationCommonMap).trim();
+
+  let textData = '### Custom Assert Documentation \n';
   const dirs = await fs.readdir(path.join('src/asserts'));
   for (const dir of dirs) {
-    textData += `\n### ${dir}\n`;
     const files = await fs.readdir(path.join('src/asserts', dir));
+    const hasAsserts = files.findIndex((file) => file.endsWith('.ts') && file !== 'index.ts');
+    if (hasAsserts === -1) continue;
+    textData += `\n#### ${dir}`;
     for (const file of files) {
       if (file.endsWith('.ts') && file !== 'index.ts') {
-        const fileData = await fs.readFile(path.join('src/asserts', dir, file));
+        const fileData = await fs.readFile(path.join('src/asserts', dir, file), { encoding: 'utf8' });
         const fileName = path.parse(file).name;
         textData += `
-        \n#### ${fileName}
+        \n##### ${fileName} (${dir})
         \n\`\`\`typescript\nimport { ${fileName} } from 'bguard/${dir}/${fileName}';\n\`\`\`
-        \n${extractJsDocContent(fileData.toString())}
-        \n
+        \n${extractJsDocContent(fileData)}
         `;
       }
     }
   }
 
-  await fs.writeFile('Asserts.md', textData);
+  textData += `\n### Contributing\nContributions are welcome! Please open an issue or submit a pull request for any bugs or feature requests.`;
+
+  await fs.writeFile('README.md', readmeStart + textData);
 };
 
 run();