+++ src/__tests__/Translation.test.ts
@@ -1,7 +1,14 @@
-import { parseOrFail, guardException, BuildSchemaError, object } from '../';
+import {
+  parseOrFail,
+  guardException,
+  BuildSchemaError,
+  object,
+  setToDefaultLocale,
+  setLocale,
+  clearLocales,
+} from '../';
 import { string } from '../asserts/string';
-import { clearLocales, setLocale, setToDefaultLocale } from '../translationMap';
-import { ExceptionContext, RequiredValidation } from '../schemas/CommonSchema';
+import { ExceptionContext, RequiredValidation } from '../commonTypes';
 
 describe('Translation', () => {
   const customEqual =
@@ -10,20 +17,23 @@ describe('Translation', () => {
       if (expected !== received) guardException(expected, received, ctx, 'somethingEqual');
     };
 
+  customEqual.key = 'somethingEqual';
+  customEqual.message = 'Something Equal';
+
   beforeEach(() => {
     clearLocales();
-    setToDefaultLocale('somethingEqual', 'Something Equal');
+    setToDefaultLocale(customEqual);
   });
 
   it('should use test translation', () => {
     setLocale('testLanguage', { somethingEqual: 'Foo is equal' });
     const testSchema = string().custom(customEqual('hello'));
-    expect(() => parseOrFail(testSchema, 'not hello', {lng: 'testLanguage'})).toThrow('Foo is equal');
+    expect(() => parseOrFail(testSchema, 'not hello', { lng: 'testLanguage' })).toThrow('Foo is equal');
   });
 
   it('should use default translation when lanugageMap is translation', () => {
     const testSchema = string().custom(customEqual('hello'));
-    expect(() => parseOrFail(testSchema, 'not hello', {lng: 'testLanguage'})).toThrow('Something Equal');
+    expect(() => parseOrFail(testSchema, 'not hello', { lng: 'testLanguage' })).toThrow('Something Equal');
   });
 
   it('should use default translation when language is not provided', () => {
@@ -32,9 +42,11 @@ describe('Translation', () => {
   });
 
   it('should fail to set same key for default translation', () => {
-    setToDefaultLocale('test:key', 'My Translation');
-    expect(() => setToDefaultLocale('test:key', 'My Translation')).toThrow('Duplicate default message key');
-    expect(() => setToDefaultLocale('test:key', 'My  Other Translation')).toThrow(BuildSchemaError);
+    customEqual.key = 'test:key';
+    customEqual.message = 'My Translation';
+    setToDefaultLocale(customEqual);
+    expect(() => setToDefaultLocale(customEqual)).toThrow('Duplicate default message key');
+    expect(() => setToDefaultLocale(customEqual)).toThrow(BuildSchemaError);
   });
 
   it('should use test translation with template reolvers', () => {
@@ -42,8 +54,13 @@ describe('Translation', () => {
       somethingEqual: 'Expected ({{e}}). Received ({{r}}). PathToError ({{p}}). Unknown ({{v}})',
     });
     const testSchema = object({ bar: string().custom(customEqual('hello')) });
-    expect(() => parseOrFail(testSchema, { bar: 'not hello' }, {lng: 'testLanguage'})).toThrow(
+    expect(() => parseOrFail(testSchema, { bar: 'not hello' }, { lng: 'testLanguage' })).toThrow(
       'Expected (hello). Received (not hello). PathToError (.bar). Unknown ({{v}})',
     );
+  });
+
+  it('should not use setLocale on "default" namespace', () => {
+    expect(() => setLocale('default', { foo: 'bar' })).toThrow('Invalid language');
+    expect(() => setLocale('default', { foo: 'bar' })).toThrow(BuildSchemaError);
   });
 });