+++ src/asserts/number/min.ts
@@ -1,23 +1,22 @@
 import { setToDefaultLocale } from '../../translationMap';
 import { guardException } from '../../exceptions';
-import type { ExceptionContext, RequiredValidation } from '../../schemas/CommonSchema';
+import { ExceptionContext, RequiredValidation } from '../../commonTypes';
 
 const minErrorMessage = 'The received value is less than expected';
 const minErrorKey = 'n:min';
 
 /**
- * @description - Asserts that a number value is not less than a specified minimum value.
- *
- * @param {number} expected - The minimum allowable value.
- * @returns {RequiredValidation} - A validation function that takes a received number and a path to the error message. Throws an error if the received value is less than the expected minimum value.
- *
+ * @description Asserts that a number value is not less than a specified minimum value.
+ * @param {number} expected The minimum allowable value.
+ * @returns {RequiredValidation} A validation function that takes a received number and an exception context.
+ * @throws {ValidationError} if the received value is less than the expected minimum value.
  * @example
  * const schema = number().custom(min(10));
  * parseOrFail(schema, 11);  // Valid
  * parseOrFail(schema, 10);  // Valid
  * parseOrFail(schema, 9);   // Throws an error: 'The received value is less than expected'
  *
- * @translation - Error Translation Key = 'n:min'
+ * @translation Error Translation Key = 'n:min'
  */
 export const min =
   (expected: number): RequiredValidation =>
@@ -27,4 +26,4 @@ export const min =
 
 min.key = minErrorKey;
 min.message = minErrorMessage;
-setToDefaultLocale(minErrorKey, minErrorMessage);
+setToDefaultLocale(min);