+++ src/asserts/string/email.ts
@@ -1,20 +1,26 @@
+import { ExceptionContext, RequiredValidation } from '../../commonTypes';
 import { guardException } from '../../exceptions';
-import type { ExceptionContext, RequiredValidation } from '../../schemas/CommonSchema';
-// TODO skloni svuda @returns jer ne daje vrednost nikakvu
-// TODOskloni ispred svako @see, @description, @param itd sve crte kao -
+import { setToDefaultLocale } from '../../translationMap';
+
 const emailRegExp = /^[^@]+@[^@]+\.[^@]+$/;
+const emailErrorMessage = 'The received value does not match the required email pattern';
+const emailErrorKey = 's:email';
 
 /**
  * @description Asserts that a string value matches the email pattern. The pattern checks for a basic email format.
- *
- * @throws an error if the received value does not match the email pattern.
- * 
+ * @returns {RequiredValidation} A validation function that takes a received string and an exception context.
+ * @throws {ValidationError} if the received value does not match the email pattern.
  * @example
  * const schema = string().custom(email());
  * parseOrFail(schema, 'example@example.com'); // Valid
  * parseOrFail(schema, 'invalid-email');      // Throws an error: 'The received value does not match the required email pattern'
+ *
+ * @translation - Error Translation Key = 's:email'
  */
 export const email = (): RequiredValidation => (received: string, ctx: ExceptionContext) => {
-  if (!emailRegExp.test(received))
-    guardException(emailRegExp, received, ctx, 'The received value does not match the required email pattern');
+  if (!emailRegExp.test(received)) guardException(emailRegExp, received, ctx, emailErrorMessage);
 };
+
+email.key = emailErrorKey;
+email.message = emailErrorMessage;
+setToDefaultLocale(email);