+++ src/index.ts
@@ -7,5 +7,5 @@ export { boolean } from './asserts/boolean';
 export { array } from './asserts/array';
 export { object } from './asserts/object';
 export { oneOfTypes } from './asserts/mix';
-export { ValidationError, BuildSchemaError, guardException} from './exceptions';
-export { setToDefaultLocale, setLocale } from './translationMap';
+export { ValidationError, BuildSchemaError, guardException } from './exceptions';
+export { setLocale, setToDefaultLocale, clearLocales } from './translationMap';