+++ CHANGELOG.md
@@ -11,3 +11,11 @@
 ### Bundle fix
 
 - Adding custom assertions to the bundle and enhancing the README file.
+
+## 0.1.0
+
+### Rename parseSchema and add translation feature
+
+- Rename parseSchema into parseOrFail
+- Added parse feature
+- Added error message translation feature